+++ app/src/main/cpp/pcsx2/pcsx2/R3000ADef.h
@@ -96,12 +96,12 @@ struct psxRegisters {
     CP2Ctrl CP2C; 		/* Cop2 control registers */
     u32 pc;				/* Program counter */
     u32 code;			/* The instruction */
-    u64 cycle;
+    u32 cycle;
     u32 interrupt;
     u32 pcWriteback;
 
     // Controls when branch tests are performed.
-    u64 iopNextEventCycle;
+    u32 iopNextEventCycle;
 
     // This value is used when the IOP execution is broken to return control to the EE.
     // (which happens when the IOP throws EE-bound interrupts).  It holds the value of
@@ -114,7 +114,7 @@ struct psxRegisters {
     s32 iopCycleEE;
     u32 iopCycleEECarry;
 
-    u64 sCycle[32];		// start cycle for signaled ints
+    u32 sCycle[32];		// start cycle for signaled ints
     s32 eCycle[32];		// cycle delta for signaled ints (sCycle + eCycle == branch cycle)
     //u32 _msflag[32];
     //u32 _smflag[32];
